+++ config/compiletime/apps/navkit_sim/StationaryGnss.hpp
@@ -4,7 +4,11 @@
 #pragma once
 
 #include "navkit/StationaryGnss.hpp"
-#include "navkit/app_support/StationaryGnssApp.hpp"
+#include "navkit/app_support/GnssEmulator.hpp"
+#include "navkit/app_support/SensorId.hpp"
+#include "navkit/app_support/SimulationApp.hpp"
+
+#include <tuple>
 
 namespace navkit::config::apps::navkit_sim
 {
@@ -12,7 +16,15 @@ namespace navkit::config::apps::navkit_sim
 struct StationaryGnssConfig
 {
     using NavKit = ::navkit::config::navkit::StationaryGnssConfig;
-    using App = ::navkit::app_support::StationaryGnssApp<StationaryGnssConfig>;
+
+    static constexpr ::navkit::app_support::SensorId PrimaryGnssSensorId =
+        NavKit::PrimaryGnssSensorId;
+    using EmulatorBindings =
+        std::tuple<::navkit::app_support::EmulatorBinding<PrimaryGnssSensorId,
+                                                          ::navkit::app_support::GnssEmulator,
+                                                          NavKit::PrimaryGnssSensor>>;
+
+    using App = ::navkit::app_support::SimulationApp<StationaryGnssConfig>;
 };
 
 } // namespace navkit::config::apps::navkit_sim