+++ include/navkit/api/config/ConfigApi.hpp
@@ -1,0 +1,10 @@
+// Copyright (c) 2026 William Gordon Carter.
+// All Rights Reserved.
+
+#pragma once
+
+#include "navkit/api/config/NavKitProductConfigPolicy.hpp"
+#include "navkit/api/config/SensorGraphConfigPolicy.hpp"
+#include "navkit/core/config/ConfigPolicy.hpp"
+#include "navkit/core/estimation/filter/FilterConfigPolicy.hpp"
+#include "navkit/core/estimation/sensor/SensorConfigPolicy.hpp"