+++ include/navkit/core/estimation/sensor/SensorId.hpp
@@ -1,0 +1,13 @@
+// Copyright (c) 2026 William Gordon Carter.
+// All Rights Reserved.
+
+#pragma once
+
+#include <cstdint>
+
+namespace navkit::core::estimation
+{
+
+using SensorId = std::uint32_t;
+
+} // namespace navkit::core::estimation